requirements.txt
PyQt6>=6.9.1
pynput>=1.7.8
mss>=10.0.0
requests>=2.32.4
Pillow>=11.2.1
betterproto>=2.0.0b7
keyboard~=0.13.5; sys_platform != 'linux'
python-xlib~=0.33; sys_platform == 'linux'